requirements.txt
werkzeug
PyJWT
Flask
Flask-SQLAlchemy
Flask-Migrate
sqlalchemy_utils
pytest
python-dotenv
pymysql
